requirements.txt
absl-py==0.9.0
pybase64==1.0.1
pycurl==7.43.0.2
PyYAML==5.3.1
